+++ src/TableGateway/AbstractTableGateway.php
@@ -198,7 +198,7 @@ abstract class AbstractTableGateway implements TableGatewayInterface
 
     /**
      * @param Select $select
-     * @return null|ResultSetInterface
+     * @return ResultSetInterface
      * @throws \RuntimeException
      */
     public function selectWith(Select $select)