+++ test/Adapter/Driver/IbmDb2/StatementIntegrationTest.php
@@ -12,6 +12,7 @@ namespace ZendTest\Db\Adapter\Driver\IbmDb2;
 use PHPUnit\Framework\TestCase;
 use Zend\Db\Adapter\Driver\IbmDb2\IbmDb2;
 use Zend\Db\Adapter\Driver\IbmDb2\Statement;
+use Zend\Db\Adapter\Exception\RuntimeException;
 
 /**
  * @group integration
@@ -74,7 +75,7 @@ class StatementIntegrationTest extends TestCase
 
         $statement = new Statement;
         $statement->initialize($db2Resource);
-        $statement->prepare("SELECT 'foo'");
+        $statement->prepare("SELECT 'foo' FROM sysibm.sysdummy1");
         $resource = $statement->getResource();
         self::assertEquals('DB2 Statement', get_resource_type($resource));
         unset($resource, $db2Resource);
@@ -101,7 +102,23 @@ class StatementIntegrationTest extends TestCase
     }
 
     /**
-     * @covers \Zend\Db\Adapter\Driver\IbmDb2\Statement::execute
+     * @covers Zend\Db\Adapter\Driver\IbmDb2\Statement::prepare
+     */
+    public function testPrepareThrowsAnExceptionOnFailure()
+    {
+        $db2Resource = db2_connect(
+            $this->variables['database'],
+            $this->variables['username'],
+            $this->variables['password']
+        );
+        $statement = new Statement;
+        $statement->initialize($db2Resource);
+        $this->setExpectedException(RuntimeException::class);
+        $statement->prepare("SELECT");
+    }
+
+    /**
+     * @covers Zend\Db\Adapter\Driver\IbmDb2\Statement::execute
      */
     public function testExecute()
     {