+++ test/Sql/Platform/IbmDb2/SelectDecoratorTest.php
@@ -26,8 +26,13 @@ class SelectDecoratorTest extends TestCase
      * @covers \Zend\Db\Sql\Platform\SqlServer\SelectDecorator::processLimitOffset
      * @dataProvider dataProvider
      */
-    public function testPrepareStatement(Select $select, $expectedPrepareSql, $expectedParams, $notUsed)
-    {
+    public function testPrepareStatement(
+        Select $select,
+        $expectedPrepareSql,
+        $expectedParams,
+        $notUsed,
+        $supportsLimitOffset
+    ) {
         $driver = $this->getMockBuilder('Zend\Db\Adapter\Driver\DriverInterface')->getMock();
         $driver->expects($this->any())->method('formatParameterName')->will($this->returnValue('?'));
 
@@ -49,6 +54,7 @@ class SelectDecoratorTest extends TestCase
 
         $selectDecorator = new SelectDecorator;
         $selectDecorator->setSubject($select);
+        $selectDecorator->setSupportsLimitOffset($supportsLimitOffset);
         $selectDecorator->prepareStatement($adapter, $statement);
 
         self::assertEquals($expectedParams, $parameterContainer->getNamedArray());
@@ -60,7 +66,7 @@ class SelectDecoratorTest extends TestCase
      * @covers \Zend\Db\Sql\Platform\IbmDb2\SelectDecorator::getSqlString
      * @dataProvider dataProvider
      */
-    public function testGetSqlString(Select $select, $ignored0, $ignored1, $expectedSql)
+    public function testGetSqlString(Select $select, $ignored0, $ignored1, $expectedSql, $supportsLimitOffset)
     {
         $parameterContainer = new ParameterContainer;
         $statement = $this->getMockBuilder('Zend\Db\Adapter\Driver\StatementInterface')->getMock();
@@ -69,6 +75,7 @@ class SelectDecoratorTest extends TestCase
 
         $selectDecorator = new SelectDecorator;
         $selectDecorator->setSubject($select);
+        $selectDecorator->setSupportsLimitOffset($supportsLimitOffset);
 
         self::assertEquals($expectedSql, @$selectDecorator->getSqlString(new IbmDb2Platform));
     }
@@ -123,12 +130,26 @@ class SelectDecoratorTest extends TestCase
         $expectedSql4 = 'SELECT * FROM ( SELECT "x".*, ROW_NUMBER() OVER () AS ZEND_DB_ROWNUM FROM "foo" "x" WHERE "x"."id" > \'10\' AND "x"."id" < \'31\' ) AS ZEND_IBMDB2_SERVER_LIMIT_OFFSET_EMULATION WHERE ZEND_IBMDB2_SERVER_LIMIT_OFFSET_EMULATION.ZEND_DB_ROWNUM BETWEEN 0 AND 5';
         // @codingStandardsIgnoreEnd
 
+        $select5 = new Select;
+        $select5->from(['x' => 'foo'])->limit(5);
+        $expectedParams5 = [];
+        $expectedPrepareSql5 = 'SELECT "x".* FROM "foo" "x" LIMIT 5';
+        $expectedSql5 = 'SELECT "x".* FROM "foo" "x" LIMIT 5';
+
+        $select6 = new Select;
+        $select6->columns([new Expression('DISTINCT(id) as id')])->from(['x' => 'foo'])->limit(5)->offset(10);
+        $expectedParams6 = [];
+        $expectedPrepareSql6 = 'SELECT DISTINCT(id) as id FROM "foo" "x" LIMIT 5 OFFSET 10';
+        $expectedSql6 = 'SELECT DISTINCT(id) as id FROM "foo" "x" LIMIT 5 OFFSET 10';
+
         return [
-            [$select0, $expectedPrepareSql0, $expectedParams0, $expectedSql0],
-            [$select1, $expectedPrepareSql1, $expectedParams1, $expectedSql1],
-            [$select2, $expectedPrepareSql2, $expectedParams2, $expectedSql2],
-            [$select3, $expectedPrepareSql3, $expectedParams3, $expectedSql3],
-            [$select4, $expectedPrepareSql4, $expectedParams4, $expectedSql4],
+            [$select0, $expectedPrepareSql0, $expectedParams0, $expectedSql0, false],
+            [$select1, $expectedPrepareSql1, $expectedParams1, $expectedSql1, false],
+            [$select2, $expectedPrepareSql2, $expectedParams2, $expectedSql2, false],
+            [$select3, $expectedPrepareSql3, $expectedParams3, $expectedSql3, false],
+            [$select4, $expectedPrepareSql4, $expectedParams4, $expectedSql4, false],
+            [$select5, $expectedPrepareSql5, $expectedParams5, $expectedSql5, true],
+            [$select6, $expectedPrepareSql6, $expectedParams6, $expectedSql6, true],
         ];
     }
 }